Feature/ci 338 view provider details

+++ src/Web/Sfa.Das.Sas.Core/Domain/Services/IGetFrameworks.cs
@@ -11,7 +11,7 @@ namespace Sfa.Das.Sas.Core.Domain.Services
 
         long GetFrameworksAmount();
 
-        long GetFrameworksOffer();
+        int GetFrameworksOffer();
 
         int GetFrameworksExpiringSoon(int daysToExpire);
     }
